Test.DocTest.Driver.CodeGen

Description

Generating DocTests from extracted information.

Documentation

data CodeGen a

Code generation monad.

Collects generated text in a GHC Doc.

Tracks the file name and the line number for current line. We emit {-# LINE ... #-} pragmas into the generated modules, but line numbers also increase by one naturally after a linebreak. If the file name does not change, and the natural line number matches the expected one, then we omit the {-# LINE ... #-} pragma to avoid cluttering the output.

type Doc = CodeGen ()

A document is code generation without a separate result.

runDoc :: [String] -> Doc -> Doc

Run the code generation and get the generated GHC Doc.

genModuleDoc :: Module -> Doc

Code generation entry-point for an extracted Module.

codeGen :: FilePath -> [Module] -> IO [FilePath]

Perform code generation for all extracted Modules, together with a Main module.

codeGenSingle :: FilePath -> Module -> IO FilePath

Perform code generation for a single extracted Module.