Test.DocTest.Driver.Extract.Ast

Description

DocTests syntax trees and extraction from source code parsed using the GHC API.

Location Information

type Loc = Either FastString RealSrcLoc

Best-effort source location.

Entities

data EntityKind

Kind of declared entities.

Constructors

Section
Haddock section heading.

DataFamily
data family.

TypeFamily
type family, open or closed.

NewType
newtype.

DataType
data.

TypeSynonym
type, typekindconstraint synonym.

Class
class, type classes.

Binding
Top-level variable or function binding.

Signature
Top-level variable or function signature.

PatternSynonym
Pattern synonym definition.

PatternSignature
Pattern synonym signature.

ForeignImport
foreign import declarations.

ForeignExport
foreign export declarations.

NamedDoc
Haddock markup, named chunk ($name).

data Entity

Declared entity information.

Constructors

Entity

Fields

kind :: !EntityKind

Entity kind.

name :: !FastString

Entity name.

declEntity :: HsDecl GhcPs -> Maybe Entity

Extract an Entity for the declaration.

This is used to determine which declaration a Haddock document should attach to.

Just e if at least one entity is found, and e is the first entity.

Nothing if there is no entity the document could attach to.

nameEntity :: String -> Entity

Create a NamedDoc entity representing a named doc comment.

DocTests

data Module

Extracted module.

Constructors

Module

Fields

filePath :: FilePath

Source file path to the original module.

modulePath :: [String]

Module name path, separated by dots (.).

importList :: [DocLine]

List of import statements to be put at the top.

topSetup :: [DocLine]

Top-level setup code, e.g., data type definitions.

testCases :: [DocTests]

Extracted test cases.

extractModule :: ParsedModule -> Module

Extract a Module from a GHC ParsedModule.

Implementation Note

The documentation in a ParsedModule is first extracted into a doc-tree (via collectDocs). By design, doctest instructions can appear as group names (Haddock markup $group-name) and as comments in verbatim code block (Haddock markup @...@). Inside this doc-tree, we inspect these places, recognise the instructions, and apply the instructions. Some instructions modify their parent scope (Captures and IOHooks bring new names in scope), so we float them above.

data DocTests

Extracted DocTest items.

Constructors

Group Entity Loc [DocTests]
Group, translated to describe.

TestProperty (NonEmpty DocLine)
Property, to be tested with QuickCheck.

TestExample (NonEmpty ExampleLine)
Example, compared with expected output.

TestMultiline (NonEmpty DocLine)
Multiline tests.

Capture CapturedContent [DocTests]
Captured string literal.

TestHook IOHook [DocTests]
IO hook: before, beforeAll, after, afterAll.

Warning Loc String
Problems detected while extracting the DocTests.

data DocLine

A single line from the DocTests, rendered with {-# LINE ... #-} and {-# COLUMN ... #-}.

Constructors

DocLine

Fields

location :: Loc

Original source location.

textLine :: String

Content of the line.

spanDocLine :: (Char -> Bool) -> DocLine -> (String, DocLine)

span, but also handles the location of DocLines.

data ExampleLine

A single line of example, accompanied by several output lines.

Constructors

ExampleLine

Fields

programLine :: DocLine

Example line (code).

expectedOutput :: [DocLine]

Output text lines.

data CapturedContent

Captured text content, made available as a variable.

Constructors

Captured

Fields

variableName :: String

Name of the variable to bring into scope.

captureMethod :: CaptureMethod

How the text content is captured.

textContent :: NonEmpty String

Text content to be captured.

data CaptureMethod

Different ways of capturing text contents.

Constructors

String
String variable.

TextStrict
Strict Text variable (from Data.Text).

TextLazy
Lazy Text variable (from Data.Text.Lazy).

ByteStringStrict
Strict ByteString variable (from Data.ByteString).

ByteStringLazy
Lazy ByteString variable (from Data.ByteString.Lazy).

ShortByteString
ShortByteString variable (from Data.ByteString.Short).

TempFile
Save as a temporary file, provide as a FilePath variable.

data IOHook

IO actions to be performed throughout the test lifetime.

Constructors

IOHook

Fields

flavour :: HookFlavour

Occasions when the IO action is expected to run.

variables :: [String]

List of variable names brought into scope by this hook.

setupCode :: NonEmpty DocLine

The code for the action.

data HookFlavour

Occasions to run an IOHook. Around and AroundAll must capture the continuation as the first parameter.

Constructors

Before
Once before each test case (result is freshly generated).

BeforeAll
Once before all test cases (result is shared).

After
Once after each test case.

AfterAll
Once after all test cases.

Around
Once around each test case (wrapper runs multiple times).

AroundAll
Once around all test cases (wrapper runs exactly once).