Test.DocTest.Driver.Utils

Description

Utility functions only used by the DocTest-driver itself during code generation. These are not used at runtime by the generated code.

Documentation

splitBy :: Eq a => a -> [a] -> [[a]]

Split by the given separator.

>>> splitBy ' ' "abc def"
["abc","def"]
>>> splitBy ' ' "abc  def"
["abc","","def"]

dosLines :: String -> [String]

lines, but with additional support for DOS line endings (CR LF).

data NaturalOrderString

String segment for natural ordering.

naturalOrdered :: String -> NaturalOrderString

Preprocess a string to be compared in the "natural order".

Natural order is lexicographical order if there is no digit character:

>>> naturalOrdered "abc" < naturalOrdered "def"
True

Natural order considers 2 < 10, while lexicographical order insists 2 > 10:

>>> naturalOrdered "image2" < naturalOrdered "image10"
True

However, using leading zeros opt into using lexicographical order.

>>> naturalOrdered "image02" > naturalOrdered "image010"
True
>>> naturalOrdered "image100" < naturalOrdered "image010"
True