Test.DocTest.FuzzySyntax

Description

Fuzzy pattern syntax used for doctest output text segments.

Documentation

type Pattern = [Segment]

Pattern string.

data Segment

Pattern segment.

Constructors

BlankLine
"<BLANKLINE>" for empty lines.

MultilineDots
"..." on its own line, matching one or more arbitrary lines.

InlineDots
"..." in the middle of line, matching one or more non-newline characters.

Plain String
Other plain string, matched verbatim.

parsePattern :: String -> Pattern

Parsing the pattern string into a Pattern structure.